Move Naga IR types from `naga/src/lib.rs` to `naga/src/ir/mod.rs`.
The types remain available at their original paths by `pub use ir::*`.
This may or may not be a temporary measure for migration.

The only changes to the moved code are:
* Fixing doc links.
* Moving `pub use block::Block` to the top.
